added 09-install-haproxy.yml playbook
initially the webapp failed when exposed through haproxy (in either
http or tcp mode).  podman logs from the howsmysalute container running
on the edge node show the following:

192.168.122.1 - - [09/Jan/2023 19:37:24] "GET /video_feed HTTP/1.1" 500
Error on request:
Traceback (most recent call last):
  File "/opt/app-root/lib64/python3.9/site-packages/werkzeug/serving.py", line 335, in run_wsgi
    execute(self.server.app)
  File "/opt/app-root/lib64/python3.9/site-packages/werkzeug/serving.py", line 324, in execute
    for data in application_iter:
  File "/opt/app-root/lib64/python3.9/site-packages/werkzeug/wsgi.py", line 500, in __next__
    return self._next()
  File "/opt/app-root/lib64/python3.9/site-packages/werkzeug/wrappers/response.py", line 50, in _iter_encoded
    for item in iterable:
  File "/app/app.py", line 13, in gen
    yield (b'--frame\r\n'
TypeError: can't concat NoneType to bytes

editing app.py in the container source and adding a comma at the end of
line 13 resolved the issue.  however, subsequent rebuilds without this
change haven't exhibited the error, but it and the fix are being
captured here just in case.